// types/seismicity.ts
export interface SeismicData {
    dyhd?: string;
    earthquakeMagnitude?: number | string;
    SS?: number | string;
    S1?: number | string;
    PGA?: number | string;
    SS_DD1?: number | string;
    S1_DD1?: number | string;
    PGA_DD1?: number | string;
    SS_DD2?: number | string;
    S1_DD2?: number | string;
    PGA_DD2?: number | string;
    SS_DD3?: number | string;
    S1_DD3?: number | string;
    PGA_DD3?: number | string;
    SS_DD4?: number | string;
    S1_DD4?: number | string;
    PGA_DD4?: number | string;
    localSoilClass?: string;
}